refact resource architect, split out ui and functional component
