+++ src/interfaces/resources/__init__.py
@@ -1,0 +1,3 @@
+from .resourcesUI import ResourcesUI
+from .resourcesControllerUI import ResourcesControllerUI
+from .resourcesAddViewUI import ResourcesAddViewUI
